Example 52b Estimator should now work
Basically uncommented the data files, changed the name to the existing data files, changed the time of the estimator to match the spectra data and also took out the fixed value for k1. This, however gives a value fairly far from the original value, so it is something to look into. At least now runs.