estimation.py for example 52a now runs
Just changed the data file to not be the "trimmed" data since that did not seem to exist. Also removed the fixed k and now get a very very different value for k. If you uncomment lines 76 and 77 then these it still runs and gives the "correct" value. Not sure why this is so far out.