Documentation update/tutorial/test updates
So this update re-arranges the folders, removes old documentation, and adds the latest documentation version. 

It also includes 3 test schedules:

test_examples.py - the original. a lot of failing problems, especially in Py3.6 and also takes almost 10 hours on my PC.

test_examples1.py - removed some of the examples that take a really long time and some that always fail. Still takes a long time (up to 4-5 hours) and still contains fails, some of which do not fail manually.

Finally,

tut_examples_test.py - this test is, I think the best way to test Kipet as it involves all of the new examples for the user, and almost certainly involves all of Kipet's capabilities and functions. It has a bit of everything and only takes an hour to run. The downside is that some of the files fail in the test, but do run manually.