New fe_factory interface and initializations in TemplateBuilder
This new branch contains a new way to call fe_factory through the use of a new module "FESimulator" which just constructs the problem for fe_factory so that the user no longer needs to re-enter the problem for fe_factory. The new method is demonstrated in 'Michaels_FESimulator.py'. Currently it is not working for the aspirin_fe_factory problems, but neither is the normal fe_factory on my PC. Both examples fail on the second finite element.

During the making of this wrapper function I encountered the issue with the inits being set to 1 and then fixed this through forcing the initializations in the TemplateBuilder to come from the initial_conditions inputted by the user. I think that this fixes the issue, however I would appreciate some feedback if this works for the other problems where this issue was experienced.