New branch to do parameter estimation using C input
The new branch includes the feature to use C as an input, doing away with the (D - SC)**2 part of the objective function and including new data tools. An example is included in the examples folder which shows how to make use of this feature. When all the new features are ironed out of it, I will update the documentation to include all the new functions on the data side as well as how to implement this. Right now the confidence intervals are not working for this problem. Once this is included and the branch updated I will do the pull request.