Estimation with concentrations only, new init for P
The latest branch adds new examples for the parameter estimation from concentration data, new data_tools for manipulating these new inputs, and a new way to initialize parameters. 

The new parameter initialization puts the initial value in the middle of the user-provided bounds, as opposed to at 1, as was done before. 

Still not working are the confidence intervals.