Concentration Input
Concentration data can now be included, along with the generation of confidence intervals for the estimated parameters. A couple new data tools were also added, along with new examples and tutorials to demonstrate to the user how to use the new inclusions. Documentation has been updated accordingly (including FESimulator, which had yet to be included)