Pre-processing options included
Latest commit includes the ability for Savitzky-Golay filter to use derivatives. With this inclusion, new bounds were set and a new check is done to set bounds on C, S, and D, depending on whether entries in the D or C matrix that are provided by the user have negative entries.

Tests were carried out successfully and new documentation is added that includes the new functions.