Lack of fit and wavelength correlation
This commit adds to the problem diagnostics arsenal of KIPET with a basic pca function that has been expanded, as well as a lack of fit assessment that can be done after the parameter estimation problem as well as a way to measure the correlation between certain wavelengths and the obtained concentration profiles. This will be incorporated into a wavelength selection tool soon.