Lack of fit and wavelength correlation final version
This is the last of the wavelength correlation and lack of fit to be added. As discussed with Larry, there is no reason to assume that the relationship between wavelength selection and lack of fit will be unimodal. So instead of wavelength minimization, we now allow the user to select the level of filter.

We provide tools that make this easier, namely, allowing the user to select certain bands of subset filter to search at what intervals, giving the user the ability to decide the subset of lambdas they wish to include.

Tutorial example is now included as well. Once the documentation is updated then a new KIPET version can be officially released with these new inclusions.