Variance estimator fixes + new data_tools
Variance estimator fixed for ipopt_lsq, negative vals in D-matrix, issues with wrong keywords and using the correct solver options. Also now able to perform the variance estimation with device variance fixed. Also suppressed the numpy least squares warning and added the option to output the time to run the vriance estimation.

Data_tools added for instruments, reading time-stamped data, making negative values to 0 or not. Also added a baseline_shift tool and a way to decrease the size of the data matrix by removing wavelengths.